Persistently track application output
The main process now stores all launched application events so the renderer process can fetch them on-demand.

This also changes all real time application events to go through IPC to simplify things. In the future it may be worth exploring using a MessagePort for this again as this solution may be somewhat inefficient.