Implement application closing
This changes the launch button into a close one once a selected application is launched. Once the close button is clicked, a kill signal will be sent to the selected application and the button will be disabled until the application fully exits.

There is currently a bug where output errors are lost once the AppLauncher component is unmounted, but it will be fixed naturally when a better method of persisting application output is implemented.